backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
diff-match-patch==20241021
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
django-import-export==4.3.9
django-jazzmin==3.0.1
django-shortuuidfield==0.1.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
uritemplate==4.2.0
urllib3==2.5.0
